bot/requirements.txt
python-telegram-bot[job-queue]==22.1
python-dotenv==1.0.0
sqlalchemy
psycopg2-binary